requirements.txt
requests
beautifulsoup4
google-genai
pypdf
python-dotenv
reportlab
rich
